[Feature] Better processors (#1378)
* Refactor pkg/plugin and isolate connector plugin (#1302)

* refactor pkg/plugin and isolate connector plugin

* fix imports

* fix tests

* fix import

* change registry to connectorRegistry

---------


* BP: WASM: Load available WASM processors (#1322)




---------


* Refactor pkg/plugin and isolate connector plugin (#1302)

* refactor pkg/plugin and isolate connector plugin

* fix imports

* fix tests

* fix import

* change registry to connectorRegistry

---------


* BP: WASM: Load available WASM processors (#1322)


---------


* Refactor processor service, processor and node (#1354)

* linter

* Run processor examples, gather specifications (#1384)

* update processor-sdk, use config parameter from conduit-commons

* add unified diff lib

* add utilities for running processor examples and collecting specifications

* update diff readme

* fix markdown linter error

* fix liter errors

* update processor-sdk

* fix race condition in test

* BP: Display processor example raw data as string (#1397)

* display raw data in processor examples as string

* ensure examples are formatted the same way in processors.json

* BP: add builtin processor part#1 (#1371)

* add processors field.set & field.subset.exclude

* delete extra test

* fix tests

* add field.rename processor

* add exclusion validation to field.rename processor

* refactor + fix error return in Process method

* add field.convert processor

* add filter processor

* only structured data

* address reviews part1

* use paramgen, update Configure method, update tests

* evaluate value for processor setField

* use a slice of referenceResolvers for processors with multiple fields params

* do not allow .Position to be set in setField processor

* update setField test, add non existent field test

* use the new Rename() method from referenceResolver

* address reviews

* address reviews2

* fix renameField bug + add New() method to each processor

* linter fix

* add processor examples

* processor json

* Update pkg/plugin/processor/builtin/convertField.go


* address reviews

* setFeild example

* setFeild example

* make generate

* fix paramgen regex

* generate

* fix regex

* add paramgen to tools

---------


* BP: Add unwrap.debezium (#1393)

* BP: Add unwrap.kafkaconnect (#1395)

* BP: Add webhook.http (#1382)

* BP: Add unwrap.opencdc (#1387)

* BP: Add encode.avro (#1401)

* updated docs, regenerate

* config

* first pass

* downgrade lib

* processing

* example

* fix config parsing

* regenerate

* fix config

* regenerate, tests

* linter, move to opencdc

* add test certs

* config refactor, tests

* more tests

* linter

* Regenerate

* fix linter

* git diff

* git diff

* regenerate

* avro examples

* run schema registry in examples on port 54321

* fix linter error

* rename processor

* fix schema registry integration test

* don't run avro examples in integration tests

---------


* BP: Add custom.javascript (#1374)

* add json.decode processor (#1405)

* add json.decode processor

* make generate

* processors

* address reviews

* make generate

* add support for decoding JSON types other than map

* make generate

* address review comments

---------


* BP: List processor plugins in API (#1406)

* update proto API definitions

* fix proto linter warnings

* deprecate ListPlugins

* implement plugin list endpoint

* processor plugin orchestrator

* move plugin endpoints under connector and processor service

* BP: add avro.decode (#1410)


* avro.decode example, remove base64 decoding

---------


* BP: Base64 decode / encode processors (#1412)

* encode

* base64 decode

* organize imports

* add encodeJson processor (#1413)

* encode json processor

* delete forgotten test

* address reviews

* linter fix

* add processor examples

* don't encode nil, output string

---------


* Builtin processors refactoring, housekeeping (#1402)

* refactor builtin processors, generate specs into separate files

* move processors into folder impl

* go mod tidy

* add example summaries

* add example summaries

* fix name of avro.decode processor

* add example summaries

* regenerate specs

* update default registry

* update default registry: fix tests, linter, regenerate

* dep upgrades

* refactorings

* move json encode processor to impl/json
* rename field.subset.exclude to field.exclude
* rename constructor for unwrap.debezium processor
* regenerate specs

* dep downgrades

---------


* fix linter

* BP: call the conditional execution methods in the processors (#1411)

* init

* simplify

* fix evaluation

* add support for batches in runnable processor

* batch test in runnable processor

* optimize case when all records are passed through in a processor

* make sure remaining records are copied over

* fix len check

* make sure processor doesn't return more records

---------


* update examples and specs

* disable goconst in processors (docs are repetitive)

* upgrade processor sdk

* add log when loading wasm processors, more docs updates

---------

+++ pkg/processor/service.go
@@ -12,33 +12,40 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:generate mockgen -destination=mock/plugin_service.go -package=mock -mock_names=PluginService=PluginService . PluginService
+
 package processor
 
 import (
 	"context"
 	"time"
 
+	sdk "github.com/conduitio/conduit-processor-sdk"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/foundation/database"
 	"github.com/conduitio/conduit/pkg/foundation/log"
 	"github.com/conduitio/conduit/pkg/foundation/metrics/measure"
 )
 
+type PluginService interface {
+	NewProcessor(ctx context.Context, pluginName string, id string) (sdk.Processor, error)
+}
+
 type Service struct {
 	logger log.CtxLogger
 
-	registry  *BuilderRegistry
+	registry  PluginService
 	instances map[string]*Instance
 	store     *Store
 }
 
-// NewService creates a new processor service.
-func NewService(logger log.CtxLogger, db database.DB, registry *BuilderRegistry) *Service {
+// NewService creates a new processor plugin service.
+func NewService(logger log.CtxLogger, db database.DB, registry PluginService) *Service {
 	return &Service{
 		logger:    logger.WithComponent("processor.Service"),
 		registry:  registry,
 		instances: make(map[string]*Instance),
-		store:     NewStore(db, registry),
+		store:     NewStore(db),
 	}
 }
 
@@ -54,7 +61,8 @@ func (s *Service) Init(ctx context.Context) error {
 	s.logger.Info(ctx).Int("count", len(s.instances)).Msg("processors initialized")
 
 	for _, i := range instances {
-		measure.ProcessorsGauge.WithValues(i.Type).Inc()
+		i.init(s.logger)
+		measure.ProcessorsGauge.WithValues(i.Plugin).Inc()
 	}
 
 	return nil
@@ -83,11 +91,29 @@ func (s *Service) Get(_ context.Context, id string) (*Instance, error) {
 	return ins, nil
 }
 
+func (s *Service) MakeRunnableProcessor(ctx context.Context, i *Instance) (*RunnableProcessor, error) {
+	if i.running {
+		return nil, ErrProcessorRunning
+	}
+
+	p, err := s.registry.NewProcessor(ctx, i.Plugin, i.ID)
+	if err != nil {
+		return nil, err
+	}
+	cond, err := newProcessorCondition(i.Condition)
+	if err != nil {
+		return nil, cerrors.Errorf("invalid condition: %w", err)
+	}
+
+	i.running = true
+	return newRunnableProcessor(p, cond, i), nil
+}
+
 // Create will create a new processor instance.
 func (s *Service) Create(
 	ctx context.Context,
 	id string,
-	procType string,
+	plugin string,
 	parent Parent,
 	cfg Config,
 	pt ProvisionType,
@@ -100,14 +126,14 @@ func (s *Service) Create(
 		cfg.Workers = 1
 	}
 
-	builder, err := s.registry.Get(procType)
+	// check if the processor plugin exists
+	p, err := s.registry.NewProcessor(ctx, plugin, id)
 	if err != nil {
-		return nil, err
+		return nil, cerrors.Errorf("could not get processor: %w", err)
 	}
-
-	p, err := builder(cfg)
+	err = p.Teardown(ctx)
 	if err != nil {
-		return nil, cerrors.Errorf("could not build processor: %w", err)
+		s.logger.Warn(ctx).Err(err).Msg("processor teardown failed")
 	}
 
 	now := time.Now()
@@ -116,12 +142,12 @@ func (s *Service) Create(
 		UpdatedAt:     now,
 		CreatedAt:     now,
 		ProvisionedBy: pt,
-		Type:          procType,
+		Plugin:        plugin,
 		Parent:        parent,
 		Config:        cfg,
-		Processor:     p,
 		Condition:     cond,
 	}
+	instance.init(s.logger)
 
 	// persist instance
 	err = s.store.Set(ctx, instance.ID, instance)
@@ -130,7 +156,7 @@ func (s *Service) Create(
 	}
 
 	s.instances[instance.ID] = instance
-	measure.ProcessorsGauge.WithValues(procType).Inc()
+	measure.ProcessorsGauge.WithValues(plugin).Inc()
 
 	return instance, nil
 }
@@ -142,15 +168,10 @@ func (s *Service) Update(ctx context.Context, id string, cfg Config) (*Instance,
 		return nil, err
 	}
 
-	// this can't really fail, this call already passed when creating the instance
-	builder, _ := s.registry.Get(instance.Type)
-
-	p, err := builder(cfg)
-	if err != nil {
-		return nil, cerrors.Errorf("could not build processor: %w", err)
+	if instance.running {
+		return nil, cerrors.Errorf("could not update processor instance (ID: %s): %w", id, ErrProcessorRunning)
 	}
 
-	instance.Processor = p
 	instance.Config = cfg
 	instance.UpdatedAt = time.Now()
 
@@ -171,13 +192,17 @@ func (s *Service) Delete(ctx context.Context, id string) error {
 		return err
 	}
 
+	if instance.running {
+		return cerrors.Errorf("could not delete processor instance (ID: %s): %w", id, ErrProcessorRunning)
+	}
+
 	err = s.store.Delete(ctx, id)
 	if err != nil {
 		return cerrors.Errorf("could not delete processor instance from store: %w", err)
 	}
 	delete(s.instances, id)
-	instance.Processor.Close()
-	measure.ProcessorsGauge.WithValues(instance.Type).Dec()
+	instance.Close()
+	measure.ProcessorsGauge.WithValues(instance.Plugin).Dec()
 
 	return nil
 }
